More work extracting the 3D canvas. Legacy canvas removed. Still work in progress.
