+++ src/main/webapp/js/components/interface/3dCanvas/TrackballControls.js
@@ -5,7 +5,9 @@
  * @author Luca Antiga 	/ http://lantiga.github.io
  */
 var THREE = window.THREE || require('three');
-THREE.TrackballControls = function ( object, domElement ) {
+THREE.TrackballControls = function ( object, domElement, viewerId ) {
+
+	this.viewerId = viewerId;
 
 	var _this = this;
 	var STATE = { NONE: - 1, ROTATE: 0, ZOOM: 1, PAN: 2, TOUCH_ROTATE: 3, TOUCH_ZOOM_PAN: 4 };
@@ -163,12 +165,12 @@ THREE.TrackballControls = function ( object, domElement ) {
 
 		var p = _this.object.position.toArray();
 
-		GEPPETTO.Console.executeImplicitCommand('G.setCameraPosition('+p[0].toFixed(places)+','+p[1].toFixed(places)+','+p[2].toFixed(places)+')');
+		GEPPETTO.Console.executeImplicitCommand(_this.viewerId+'.setCameraPosition('+p[0].toFixed(places)+','+p[1].toFixed(places)+','+p[2].toFixed(places)+')');
 
 		var u = _this.object.rotation.toArray();
 		var l = _eye.length();
 
-		GEPPETTO.Console.executeImplicitCommand('G.setCameraRotation('+u[0].toFixed(places)+','+u[1].toFixed(places)+','+u[2].toFixed(places)+','+l.toFixed(places)+')');
+		GEPPETTO.Console.executeImplicitCommand(_this.viewerId+'.setCameraRotation('+u[0].toFixed(places)+','+u[1].toFixed(places)+','+u[2].toFixed(places)+','+l.toFixed(places)+')');
 
 		_this.cameraByConsoleLock = true;
 		_this.cameraChanged = false;
