+++ src/main/webapp/js/geppettoModel/ModelFactory.js
@@ -442,8 +442,6 @@ define(function (require) {
              * Creates and populates initial instance tree skeleton with any instance that needs to be visualized
              */
             createInstances: function (geppettoModel) {
-                // reset scene complexity index
-                GEPPETTO.SceneController.reset();
 
                 var instances = [];
 
@@ -911,10 +909,7 @@ define(function (require) {
                                     instances[j].setVisualGroups(visualType.getVisualGroups());
                                 }
 
-                                // increase scene complexity counter
-                                if (visualType.getMetaType() == GEPPETTO.Resources.COMPOSITE_VISUAL_TYPE_NODE) {
-                                    GEPPETTO.SceneController.complexity += visualType.getVariables().length;
-                                }
+
                             }
                         }
 
@@ -1293,11 +1288,6 @@ define(function (require) {
                                     explodedInstance.extendApi(AVisualGroupCapability);
                                     explodedInstance.setVisualGroups(visualType.getVisualGroups());
                                 }
-
-                                // increase scene complexity counter
-                                if (visualType.getMetaType() == GEPPETTO.Resources.COMPOSITE_VISUAL_TYPE_NODE) {
-                                    GEPPETTO.SceneController.complexity += visualType.getVariables().length;
-                                }
                             }
 
                             // check if it has connections and inject AConnectionCapability
@@ -1371,10 +1361,6 @@ define(function (require) {
                                 newlyCreatedInstance.setVisualGroups(visualType.getVisualGroups());
                             }
 
-                            // increase scene complexity counter
-                            if (visualType.getMetaType() == GEPPETTO.Resources.COMPOSITE_VISUAL_TYPE_NODE) {
-                                GEPPETTO.SceneController.complexity += visualType.getVariables().length;
-                            }
                         }
 
                         // check if it has connections and inject AConnectionCapability
