+++ src/main/webapp/js/pages/geppetto/GEPPETTO.Main.js
@@ -6,8 +6,8 @@
  * @author giovanni@openworm.org (Giovanni Idili)
  * @author  Jesus R. Martinez (jesus@metacell.us)
  */
-define(function(require) {
-    return function(GEPPETTO) {
+define(function (require) {
+    return function (GEPPETTO) {
 
         var $ = require('jquery'),
             React = require('react'),
@@ -20,26 +20,16 @@ define(function(require) {
          */
         GEPPETTO.Main = {
 
-            StatusEnum: {
-                DEFAULT: 0,
-                CONTROLLING: 1,
-                OBSERVING: 2
-            },
-
             idleTime: 0,
             disconnected: false,
             status: 0,
             statusWorker: null,
 
-            getVisitorStatus: function() {
-                return this.status;
-            },
-
-            getStatusWorker: function() {
+            getStatusWorker: function () {
                 return this.statusWorker;
             },
 
-            startStatusWorker: function() {
+            startStatusWorker: function () {
                 //create web worker for checking status
                 if (this.statusWorker != undefined) {
                     this.statusWorker.terminate();
@@ -49,7 +39,7 @@ define(function(require) {
                 this.statusWorker.postMessage(2000);
 
                 //receives message from web worker
-                this.statusWorker.onmessage = function(event) {
+                this.statusWorker.onmessage = function (event) {
                     if (window.Project != null || undefined) {
                         var experiments = window.Project.getExperiments();
                         var pull = false;
@@ -71,7 +61,7 @@ define(function(require) {
             /**
              * Initialize web socket communication
              */
-            init: function() {
+            init: function () {
                 GEPPETTO.MessageSocket.connect(GEPPETTO.MessageSocket.protocol + window.location.host + '/' + window.BUNDLE_CONTEXT_PATH + '/GeppettoServlet');
                 GEPPETTO.Console.debugLog(GEPPETTO.Resources.GEPPETTO_INITIALIZED);
             },
@@ -79,7 +69,7 @@ define(function(require) {
             /**
              * Idle check
              */
-            idleCheck: function() {
+            idleCheck: function () {
                 if (GEPPETTO.Main.idleTime > -1) {
                     var allowedTime = 2, timeOut = 4;
                     if (!GEPPETTO.Main.disconnected) {
@@ -94,7 +84,7 @@ define(function(require) {
                             $('#infomodal-text').html(GEPPETTO.Resources.IDLE_MESSAGE);
                             $('#infomodal-btn').html("Yes");
 
-                            $('#infomodal-btn').html("Yes").click(function() {
+                            $('#infomodal-btn').html("Yes").click(function () {
                                 $('#infomodal').modal('hide');
                                 GEPPETTO.Main.idleTime = 0;
 
@@ -123,7 +113,7 @@ define(function(require) {
                             GEPPETTO.MessageSocket.close();
 
                             var webGLStarted = GEPPETTO.init(GEPPETTO.FE.createContainer());
-                            var webWorkersSupported = (typeof(Worker) !== "undefined") ? true : false;
+                            var webWorkersSupported = (typeof (Worker) !== "undefined") ? true : false;
 
                             if (!webGLStarted || !webWorkersSupported) {
                                 GEPPETTO.FE.notifyInitErrors(webGLStarted, webWorkersSupported);
@@ -140,13 +130,11 @@ define(function(require) {
         // Application logic.
         // ============================================================================
 
-        $(document).ready(function() {
-            //Create canvas
-            var webGLStarted = GEPPETTO.webGLAvailable();
-            var webWorkersSupported = (typeof(Worker) !== "undefined") ? true : false;
+        $(document).ready(function () {
+            var webWorkersSupported = (typeof (Worker) !== "undefined") ? true : false;
 
             //make sure webgl started correctly
-            if (!webGLStarted || !webWorkersSupported) {
+            if (!webWorkersSupported) {
                 GEPPETTO.FE.notifyInitErrors(webGLStarted, webWorkersSupported);
             } else {
                 GEPPETTO.FE.initialEvents();
@@ -155,12 +143,12 @@ define(function(require) {
                 setInterval(GEPPETTO.Main.idleCheck, 240000); // 1 minute
                 var here = $(this);
                 //Zero the idle timer on mouse movement.
-                here.mousemove(function(e) {
+                here.mousemove(function (e) {
                     if (GEPPETTO.Main.idleTime > -1) {
                         GEPPETTO.Main.idleTime = 0;
                     }
                 });
-                here.keypress(function(e) {
+                here.keypress(function (e) {
                     if (GEPPETTO.Main.idleTime > -1) {
                         GEPPETTO.Main.idleTime = 0;
                     }
@@ -170,7 +158,7 @@ define(function(require) {
                 GEPPETTO.Main.init();
 
                 var visibleExperiments = false;
-                $('#experimentsButton').click(function(e) {
+                $('#experimentsButton').click(function (e) {
                     if (!visibleExperiments) {
                         $('#console').hide();
                         $("#pythonConsole").hide();
@@ -185,7 +173,7 @@ define(function(require) {
                     }
                 });
 
-                $('#consoleButton').click(function(e) {
+                $('#consoleButton').click(function (e) {
                     $('#console').show();
                     $('#experiments').hide();
                     $("#pythonConsole").hide();
@@ -193,7 +181,7 @@ define(function(require) {
                     visibleExperiments = false;
                 });
 
-                $('#pythonConsoleButton').click(function(e) {
+                $('#pythonConsoleButton').click(function (e) {
                     $('#console').hide();
                     $('#experiments').hide();
                     $("#pythonConsole").show();
