+++ src/main/webapp/js/pages/geppetto/geppetto.ejs
@@ -42,7 +42,6 @@ $content
     <div id="ControlsMenuButton" class="row"></div>
     <div id="sim-toolbar" class="row"></div>
     <div id="tutorial"></div>
-    <div id="camera-controls"></div>
     <div id="foreground-toolbar" class="row"></div>
     <div id="github">
         <a href="https://github.com/openworm/org.geppetto" target="_blank">
